+++ R/feature_representation.R
@@ -375,8 +375,7 @@ methods::setMethod("feature_representation",
       inherits(solution, "Raster"),
       number_of_zones(x) == raster::nlayers(solution),
       raster::compareCRS(x$data$cost@crs, solution@crs),
-      raster::compareRaster(x$data$cost, solution[[1]], res = TRUE,
-        tolerance = 1e-5, stopiffalse = FALSE),
+      is_comparable_raster(x$data$cost, solution[[1]]),
       min(raster::cellStats(solution, "min")) >= 0,
       max(raster::cellStats(solution, "max")) <= 1)
     # subset planning units with finite cost values